Sniff: move `getFQClassNameFromNewToken()` method to `ResolveHelper`
Along the same lines as for other utility functions, the `getFQClassNameFromNewToken()` method is only needed by a few sniffs so moving the method out of the `Sniff` class and into the `ResolveHelper` class is appropriate.

This is a preparation step to be able to switch to using the upcoming PHPCSUtils sniff abstract classes in the future.

Includes changing the method to be `static`.

Includes changing all relevant function calls to the method and moving (and adjusting) the unit tests.