New `MiscHelper` class for miscellaneous utility methods
Move functionality which is only needed by select sniffs out of the base `Sniff` class.
This is a preparation step to be able to switch to using the upcoming PHPCSUtils sniff abstract classes in the future.

Also note that it is expected that the `isUseOfGlobalConstant()` method will at some point in the future be replaced by functionality expected to be included in a future PHPCSUtils version.

As part of this move:
* The method has been made `static` and
* The method is now isolated/stand-alone.

Includes changing all relevant function calls to these methods and moving (and adjusting) the unit tests.