FunctionNameRestrictions/RemovedPHP4StyleConstructors: bow out when testVersion min >= 8.0
As of PHP 8.0, PHP 4-style constructors are treated as normal methods no matter what.

If the `testVersion` set by the end user has a minimum supported PHP version of PHP 8.0 or higher, the sniff should stay silent and presume that the method with the same name as the class is not a PHP 4 constructor to begin with.

Tested via the existing test cases.