AbstractInitialValueSniff: prevent potential infinite loop
The defensive coding which is in place for a few "shouldn't be possible" cases, does not move the stackPtr forward, which means that if those conditions are hit, there is the potential for the sniff getting into an infinite loop.

While this shouldn't be possible with valid PHP code, the infinite loop should still be avoided.

Fixed now.

Includes tests via the `NewHeredoc` sniff (which already has a `@covers` tag for the `AbstractInitialValueSniff`).