CONTRIBUTING: update info about running PHPCS on this repo
Follow up on PR 1352, which dropped support for PHPCS < 3.7.1.

As the `PHPCSDevCS` repo is now a direct dependency and the work-arounds we had in place for it have been removed in commit 17e7a63d53e94c5cdabf6788ea520bcebe166607, the information about these work-arounds can also be removed from the `CONTRIBUTING.md` guide.