Interfaces/InternalInterfaces: detect interfaces extending select internal interfaces
Interfaces can `extend` other interfaces.

I have verified whether this is allowed for the internal interfaces and turns out (unsurprisingly) that this is not allowed for the `DateTimeInterface`.
It is fine for the other internal interfaces though.

This commit updates the sniff to take this into account and trigger an error when the `DateTimeInterface` is being extended by a userland interface declaration.

Includes unit tests.