Interfaces/RemovedSerializable: detect PHP 8.1+ enums implementing Serializable
Enums can implement interfaces, but they do not support the magic `__[un]serialize()` methods.

This commit ensures that enums implementing the deprecated `Serializable` interface will be flagged with a dedicated error message which doesn't reference the solution which is available for classes.

Includes tests.

Includes removing a condition which is redundant as this is already checked via the `ObjectDeclarations::findImplementedInterfaceNames()` method.