Classes/NewClasses: improve performance
This sniff would try to get the "properties" of **all** variables, while only OO property declarations have "properties".

This meant that in the majority of cases, the function call to `Variables::getMemberProperties()` would result in an exception which would need to be created, then caught and discarded.

By doing a preliminary check on whether a `T_VARIABLE` is a property or not and bowing out early in those cases, we prevent the need for and the overhead of all those exceptions, which should make a significant difference in performance.

Note: this does mean the `Scopes::isOOProperty()` method will be called twice, once here and once within the `Variables::getMemberProperties()` method, but the overhead of that duplicate function call is not comparable with the overhead of the exception creation.

Related to 1477