Composer: update the PHPUnit Polyfills
A PHPUnit 10 compatible version of the PHPUnit Polyfills has been released, so let's start using it.

Note: by rights, the `phpunit/phpunit` requirement should be removed in favour of deferring to the PHPUnit Polyfills to handle the supported versions, however as we have a few specific requirements, I'm leaving it in for now.

As a reminder of why these specific version constraints are in place:
* The `4.8.36` and `5.7.21` minimums are to be line with the PHPUnit Polyfills.
* The `9.3.4` is to allow for warming the cache for code coverage to prevent PHP Parser from interferring.
* The `10.1.0` is related to all the config changes in PHPUnit itself and makes sure that the (upcoming) separate PHPUnit 10 config file will validate.

Ref:
* https://github.com/Yoast/PHPUnit-Polyfills/releases/tag/2.0.0