ParameterValues/NewIconvMbstringCharsetDefault: fix false negative
If a (named) parameter is passed without a value, the sniff would not throw an error, while it should.

Even though it is a parse error, the sniff can safely conclude that the target array keys will not be set if there is no value for the parameter.

Includes test.

Co-authored-by: name Diede Exterkate <diede@yoast.com>