ParameterValues/NewIconvMbstringCharsetDefault: bug fix - sniff gets confused over list param
The `iconv_mime_encode()` function expects an array as the third `$options` parameter, but the sniff did not verify that a short array token was for a short array, not a short list.

This meant that the sniff would incorrectly treat short lists as if they were arrays, leading to potential false positives + false negatives.
Even though that is silly code and shouldn't exist in real life, the sniff should still handle it correctly.

Fixed now.

Includes test.

Co-authored-by: name Diede Exterkate <diede@yoast.com>