ParameterValues/NewIconvMbstringCharsetDefault: handle PHP 7.4+ spread in array
PHP 7.4+ introduced "spread in array" (unpacking of one array into another).

As this sniff looks for specific keys in a hard-coded array passed to the `iconv_mime_encode()` function, this could confuse things and could lead to false positives for keys which are being set via the spread out array.

Along the same lines as the `Undetermined` warning when the value of the `$options` parameter is variable, this commit changes the errors for missing array keys in a found `$options` arrays to warnings when the sniff also detected array unpacking within the same array parameter.

Includes tests.

Co-authored-by: name Diede Exterkate <diede@yoast.com>