PHP 8.3 | :sparkles: New `PHPCompatibility.ParameterValues.RemovedLdapConnectSignatures` sniff (RFC)
> - LDAP
>   . Calling ldap_connect() with separate hostname and port is deprecated.

Refs:
* https://wiki.php.net/rfc/deprecations_php_8_3#deprecate_calling_ldap_connect_with_2_parameters
* https://github.com/php/php-src/blob/548fc6a8185625bf50c708a9337db01f14860ba1/UPGRADING#L146-L148
* php/php-src 5177
* https://github.com/php/php-src/commit/69a8b63ecf4fec1b35ef4da1ac9579321c45f97f

This adds a new sniff to detect the above.

**Note**:
Normally this kind of change would be handled via the `RemovedFunctionParameters` sniff, but this function has *three* distinct signatures and the signatures are being deprecated in reverse order, i.e. the 2-param signature is being deprecated in PHP 8.3, while the 3+ param signature is being deprecated in PHP 8.4.
For that reason, the `RemovedFunctionParameters` sniff is not suitable to handle this.

Also note that, while the three+ param signature is still supported, passing a separate `$host` and `$port` is also deprecated for that signature. Instead a `$uri` should be passed and `$port` should be set to `null`.
The sniff does take this into account and handles this via a separate error code.

Includes tests and documentation.

Related to 1589