Keywords/ForbiddenNames: bug fix for method names
Brought to my attention during WordCamp The Netherlands by tfrommen.

When a keyword only became reserved in PHP 7.0 or later, it doesn't matter whether the code needs to run on PHP < 7.0 or not, when determining whether to flag a method declaration.

For example:
* `match` became a reserved keyword in PHP 8.0.
* Method names can be a reserved keyword since PHP 7.0.
* The code needs to run on PHP 5.5 and higher.

Given the above, a method declared with the name `match` should not be flagged, as in PHP < 8.0 it was not a reserved keyword yet, so the fact that reserved keywords cannot be used for methods names in PHP < 7.0 is irrelevant.

This commit fixes this issue.

Note: this was previously not an issue as PHP didn't introduce new (single word) hard reserved keywords between PHP 7.0 - 7.3.

Includes updated tests.