Keywords/ForbiddenNames: bug fix for OO constant names
Follow up on 1633.

The same as applies to method names also applies to OO constant names:
> When a keyword only became reserved in PHP 7.0 or later, it doesn't matter whether the code needs to run on PHP < 7.0 or not, when determining whether to flag an OO constant declaration.

For example:
* `match` became a reserved keyword in PHP 8.0.
* OO constant names can be a reserved keyword since PHP 7.0.
* The code needs to run on PHP 5.5 and higher.

Given the above, an OO constant declared with the name `match` should not be flagged, as in PHP < 8.0 it was not a reserved keyword yet, so the fact that reserved keywords cannot be used for methods names in PHP < 7.0 is irrelevant.

This commit fixes this issue.

Note: this was previously not an issue as PHP didn't introduce new (single word) hard reserved keywords between PHP 7.0 - 7.3.

Includes updated tests.

Includes minor touch up of the changes made in 1633 (comment and consistent return in callbacks in the test generation).