Keywords/ForbiddenNames: bug fix for namespace names
Follow up on 1633.

A similar exception as applies to method names also applies to namespace names.

When a keyword only became reserved in PHP 8.0 or later, it doesn't matter whether the code needs to run on PHP < 8.0 or not, when determining whether to flag a namespace declaration.

For example:
* `match` became a reserved keyword in PHP 8.0.
* Namespace names can contain reserved keywords since PHP 8.0.
* The code needs to run on PHP 7.3 and higher.

Given the above, a namespace declared with the name `match` should not be flagged, as in PHP < 8.0 it was not a reserved keyword yet, so the fact that reserved keywords cannot be used for namespace names in PHP < 8.0 is irrelevant.

This commit fixes this issue.

Note: this was previously not an issue as PHP didn't introduce new (single word) hard reserved keywords between PHP 7.0 - 7.3.

Includes updated tests.