FunctionDeclarations/RemovedOptionalBeforeRequiredParam: add extra tests/constant expressions
Add some extra tests with `null` in the default value, but not as a stand-alone value, but as part of a constant expression.

PHP itself will only evaluate constant expressions at runtime (when the function is called), not at compile time (when the deprecation is flagged).
In practice, this means that a default value containing a constant expression which evaluates to `null` will always be flagged as deprecated by PHP if the parameter is before a required one and that the exception for typed properties with a `null` default does not apply to constant expressions.

This means that the sniffs should also disregard any `null` token within a constant expression, as the only thing that's relevant is that case, is that there is a default value.
Given that, the sniff is handling this correctly as-is.

Also see: https://github.com/php/php-src/issues/13752