FunctionDeclarations/RemovedOptionalBeforeRequiredParam: flag nullable types / PHP 8.1
The [recent RFC proposing to deprecate implicit nullable types](https://wiki.php.net/rfc/deprecate-implicitly-nullable-types) brough an oversight in the `PHPCompatibility.FunctionDeclarations.RemovedOptionalBeforeRequiredParam` sniff to my attention.

To quote [the RFC](https://wiki.php.net/rfc/deprecate-implicitly-nullable-types):
> Even though signatures which contain an optional parameter before a required one were deprecated in PHP 8.0, the case of implicitly nullable types was left alone at that time due to BC concerns. This exclusion caused some bugs in the detection of which signatures should emit the deprecation notice. Indeed, the following signature only emits a deprecation as of PHP 8.1:
>
> `function bar(T1 $a, ?T2 $b = null, T3 $c) {}`

At the time of writing this sniff this additional deprecation notice was not yet in place and was therefore not taken into account.

This commit updates the sniff to _also_ throw deprecation notices for optional parameters before required ones, where the default value is `null` and the type is nullable.

Includes unit tests safeguarding the fix.

Ref: https://github.com/php/php-src/commit/c939bd2f10b41bced49eb5bf12d48c3cf64f984a