FunctionDeclarations/RemovedOptionalBeforeRequiredParam: flag union types with null / PHP 8.3
The [recent RFC proposing to deprecate implicit nullable types](https://wiki.php.net/rfc/deprecate-implicitly-nullable-types) brough an oversight in the `PHPCompatibility.FunctionDeclarations.RemovedOptionalBeforeRequiredParam` sniff to my attention.

To quote [the RFC](https://wiki.php.net/rfc/deprecate-implicitly-nullable-types):
> Even though signatures which contain an optional parameter before a required one were deprecated in PHP 8.0, the case of implicitly nullable types was left alone at that time due to BC concerns. This exclusion caused some bugs in the detection of which signatures should emit the deprecation notice.
> <snip>
> And the signature that uses the generalized union type signature:
>
> `function test(T1 $a, T2|null $b = null, T3 $c) {}`
>
> only emits the deprecation notice properly as of PHP 8.3.

At the time of writing this sniff this additional deprecation notice was not yet in place and was therefore not taken into account.

This commit updates the sniff to _also_ throw deprecation notices for optional parameters before required ones, where the default value is `null` and the type is `null` or a union/DNF type which includes `null`.

Includes unit tests safeguarding the fix.

Refs:
* php/php-src 11488
* php/php-src 11497
* https://github.com/php/php-src/commit/68ef3938f42aefa3881c268b12b3c0f1ecc5888d