PHP 8.4 | :sparkles: New `PHPCompatibility.FunctionDeclarations.RemovedImplicitlyNullableParam` sniff (RFC)
> - Core:
>   . Implicitly nullable parameter types are now deprecated.
>     RFC: https://wiki.php.net/rfc/deprecate-implicitly-nullable-types

As this deprecation looks to be high frequency - 880 packages out of the top 2000 Composer packages affected and that's just the tip of the iceberg -, I figured it would be good to have a sniff available early.

The fix for the issue is relatively easy and there are multiple options available, though some could result in BC-breaks for the package making the fix.

Includes tests and documentation.

Refs:
* https://wiki.php.net/rfc/deprecate-implicitly-nullable-types
* https://github.com/php/php-src/blob/330cc5cdb2096c5d41041eaae1f1cc0ed7e36898/UPGRADING#L271C1-L273C70
* php/php-src 12959
* https://github.com/php/php-src/commit/330cc5cdb2096c5d41041eaae1f1cc0ed7e36898

Partially based on PR 1689

Co-authored-by: Dan Wallis <dan@wallis.nz>