FunctionDeclarations/RemovedOptionalBeforeRequiredParam: add tests with PHP 8.1 new in initializers
PHP 8.1 introduced the ability to use `new` in initializers. This now opens up the possibility that `null` is passed to an initializer.

In practice though, this has no impact on the sniff as the "`new` object" is a default value making the parameter optional, so should still be flagged when used before a required parameter.