Classes/NewTypedProperties: bug fix
Follow up on 1498, which added a test and concluded the edge case was already handled correctly, while it was not.

Basically, it comes down to the sniff trying to skip to the end of a multi-property statement, but not taking PHP close tags into account correctly, meaning too much code was being skipped.

It's unlikely this bug will have been seen in the wild, but the sniff should still handle things correctly.

Fixed now, including a test safeguarding the fix.