ParameterValues/RemovedMbStrrposEncodingThirdParam: improve message precision
Throw the error on the first non-empty token of the parameter which is being flagged, instead of on the first token, which, in most cases, would be whitespace following the comma separating the parameter from the previous token.
This will more clearly indicate the issue when people use the PHPCS `code` view, especially for multi-line function calls.