Attributes/NewAttributes: split off flagging of multiline attributes
Attributes have been implemented in a manner which makes them largely cross-version compatible (no parse error) with older PHP versions, however, there are a few exceptions.

Most notably, the following usages of attributes are not cross-version compatible:
- Multi-line attributes.
- Inline attributes, i.e. an attribute followed by other code on the same line.
- Attributes containing (something which looks like) a PHP close tag in a string argument passed to the attribute on the same line as the attribute opener.

Other than that, attributes are cross-version compatible, though (obviously), the functionality attached to the attribute will not work when the code is run on PHP < 8.0.

Until now, the `NewAttributes` sniff did not distinguish between these four situations, which made the sniff annoyingly noisy.

This commit is the first step towards changing this and introduces a separate error (and error code) for multi-line attributes, which are not cross-version compatible.

The change is covered by the pre-existing test cases, some of which will now throw the new error as well.

Related to 1480