Attributes/NewAttributes: split off flagging of inline attributes
Attributes have been implemented in a manner which makes them largely cross-version compatible (no parse error) with older PHP versions, however, there are a few exceptions.

Most notably, the following usages of attributes are not cross-version compatible:
- Multi-line attributes.
- Inline attributes, i.e. an attribute followed by other code on the same line.
- Attributes containing (something which looks like) a PHP close tag in a string argument passed to the attribute on the same line as the attribute opener.

Other than that, attributes are cross-version compatible, though (obviously), the functionality attached to the attribute will not work when the code is run on PHP < 8.0.

Until now, the `NewAttributes` sniff did not distinguish between these four situations, which made the sniff annoyingly noisy.

This commit is the second step towards changing this and introduces a separate error (and error code) for inline attributes, which are not cross-version compatible.

Notes:
* This new error will not flag multiple attributes on one line. While yes, an attribute after an attribute closer is "code", it is code only on PHP 8.0, where this wouldn't be a problem.
* If an attribute is both multi-line as well as inline (code after), both errors should be thrown, as hiding one error behind another is bad practice for sniffs.
* In contrast to the other errors, this error will not be thrown on the attribute _opener_, but on the attribute _closer_, to make it more straight forward to see the issue in the code view (as the code will be directly after the attribute closer.

The change is covered by the pre-existing test cases, some of which will now throw the new error as well.
Additionally, a new test case is introduced to verify that when an attribute is both inline as well as multi-line, both errors will be thrown.

Related to 1480