Attributes/NewAttributes: add specific error for close tag in text string param
Attributes have been implemented in a manner which makes them largely cross-version compatible (no parse error) with older PHP versions, however, there are a few exceptions.

Most notably, the following usages of attributes are not cross-version compatible:
- Multi-line attributes.
- Inline attributes, i.e. an attribute followed by other code on the same line.
- Attributes containing (something which looks like) a PHP close tag in a string argument passed to the attribute on the same line as the attribute opener.

Other than that, attributes are cross-version compatible, though (obviously), the functionality attached to the attribute will not work when the code is run on PHP < 8.0 (where the attribute will just be ignored and seen as a comment).

Until now, the `NewAttributes` sniff did not distinguish between these four situations, which made the sniff annoyingly noisy.

This commit is the third step towards changing this and introduces a separate error (and error code) for attributes containing a text string which contains text which looks like a PHP close tag, when the text string is on the same line as the attribute opener.

Example of the problematic behaviour: https://3v4l.org/oarGV

Notes:
* This third error is will not cause parse errors and may easily go unnoticed, but can cause a security issue (code exposure) and broken functionality.
* Also note that text string parameters which is not on the same line as the attribute opener are not flagged as those have the parse error problem.
    The code exposure problem only comes into play once the parse error has been fixed as until then, the text string containing the `?>` is not seen as part of the comment in PHP < 8.0, while the incompatibility problem lies in PHP respecting close tags within comment and being able to jump out of PHP when those are encountered.

The change is covered by pre-existing test cases, one of which will now (also) throw the new error.
A small tweak has been made to the test case to allow for safeguarding properly that the sniff will throw both the `FoundMultiLine`, `FoundInline`, as well as the `FoundCloseTag` error and doesn't hide one behind the other.

Related to 1480