Attributes/NewAttributes: downgrade the "generic" attributes error to a warning
Now each of the situations which cause real PHP compatibility issues are flagged as errors with their own error code, the generic "attribute found" error can be downgraded to a warning as it will only flag attributes which are cross-version compatible, though there may still be loss of functionality.

Includes a change to the message text to make it more informative.

Related to 1480