Attributes/NewAttributes: more modular error codes for the warning
The `Found` warning is still very generic and would be thrown for every cross-version compatible attribute, which can make it very noisy and make the temptation to turn it off large, while at times, the warning can prevent bugs from slipping in due to it flagging functionality which will not be executed on PHP < 8.0.

So this commit attempts to make the error code a little more modular to allow for more selectively excluding the error code.

Considerations which were taken into account:
* If the error code was made _fully_ modular (`[AttributeName]Found`), it would mean that for (valid uses of) each attribute a separate ruleset exclusion would need to be added, which would make the sniff very fiddly.
* At the same time, just having the `Found` error and excluding that, would, as pointed out above, quickly ignore too much.

Ideally, we would be able to group all attributes by "source application", but at this time, that would cause a huge maintenance overhead.
Once namespace and import use statement resolution is added to the sniff (once available in PHPCSUtils), this will become easier though as groups (and error codes!) could be defined based on the namespace "prefix".

With this last bit in mind, I've currently implemented the modularization only for the most commonly adopted attribute classes.

I have looked at which applications offer attributes and whether these are (semi-)widely adopted or are expected to be widely adopted soonish, basing this in part on the Exakat articles and in part on publicly available information about dependencies from GitHub and Packagist.

For those attribute "groups", separate error codes have now been introduced which allow for selectively excluding the warning for a whole group of attributes.

Once namespace and import use statement resolution is added to the sniff, this code can be simplified a lot (no more need for the attribute class name lists, just base the error code on the namespace prefix in the FQN).
Additionally, once namespace and import use statement resolution is added to the sniff, further modularization based on namespace prefix can be added to the sniff.

This is also the reason not make the warnings remaining in the `Found` code more modular at this time as the above mentioned future change would then invalidate all previously made excludes/ignore annotations and be a breaking change in itself, while the partially modularization done now allows more easily for expanding this in the future without necessarily breaking existing excludes.

Notes:
* Even though the Exakat articles mention ShopWare, Tuleap and Contao as other common sources of attribute, based on the same articles and a dependency check, no lists/condition for these have been added as they are mostly, if not exclusively, used within their own application and not adopted by the wider field of PHP applications.

Includes tests for the additional error codes, though this needs to be verified manually until the test framework would be adjusted to test error codes.

Related to 1480

Refs:
* https://www.php.net/manual/en/reserved.attributes.php
* https://docs.phpunit.de/en/main/attributes.html
* https://github.com/JetBrains/phpstorm-attributes
* https://www.doctrine-project.org/projects/doctrine-orm/en/3.2/reference/attributes-reference.html
* https://www.exakat.io/en/adoption-of-php-8-attributes-in-2022/
* https://www.exakat.io/en/php-8-attribute-usage-in-2023/