Attributes/NewAttributes: move "close tag in text string" tests to separate file
These tests have to be the last tests in a file, so having them in the main file, makes updating the tests fiddly as you need to keep updating the line numbers for the tests related to the "close tag in text string" errors.

This commit splits the test case file into two files to make adding new tests to the main file more straight forward.