PHP 8.4 | NewClasses: detect use of new StreamBucket class (RFC)
> . stream_bucket_make_writeable() and stream_bucket_new() will now return a
>   StreamBucket instance instead of an stdClass instance.
>   RFC: https://wiki.php.net/rfc/dedicated_stream_bucket

Note: this commit only adds detection of referencing of the new class (as a type declaration and such).

It does **not** check usage of the above mentioned functions, as the return value is still an object and has the same properties (and one extra) as the previous `stdClass` object, so code calling those functions should still work without any cross-version compatibility issues.

Also note that the RFC includes a proposal to deprecate the `$datalen` property in the return value in favour of the (new) `$dataLength` property.
This proposal, however, is slated for the _next_ PHP version (8.5 or 9.0), not for PHP 8.4.

Refs:
* https://wiki.php.net/rfc/dedicated_stream_bucket
* https://github.com/php/php-src/blob/c42615782334323511cda18a8f0dd3dc19ed6256/UPGRADING#L361-L363
* php/php-src 13111
* https://github.com/php/php-src/commit/be2f454d6ebcaf6f58c41544f21f53a5109b1e5d

Related to 1731