PHP 8.4 | :sparkles: New PHPCompatibility.ParameterValues.RemovedXmlSetHandlerCallbackUnset sniff (RFC)
> . Passing non-callable strings to the xml_set_*_handler() functions is now
>   deprecated.
>   RFC: https://wiki.php.net/rfc/deprecations_php_8_4#xml_set_object_and_xml_set_handler_with_string_method_names

While PHPCS is not particularly suitable to validate whether a string is callable or not, the _"This would also mean to unset a handler the value of null must be used instead of an empty string which is also currently allowed."_ part of the RFC should be sniffable.

This commit introduces a new sniff to detect and flag this.

Includes tests.
Includes documentation.

Refs:
* https://wiki.php.net/rfc/deprecations_php_8_4#xml_set_object_and_xml_set_handler_with_string_method_names
* https://github.com/php/php-src/blob/c42615782334323511cda18a8f0dd3dc19ed6256/UPGRADING#L496-L497
* php/php-src 15293
* https://github.com/php/php-src/commit/25b4696530c5dfad6b04a57c274200beec51ab0d

Related to 1731