PHP 8.4 | OptionalToRequiredFunctionParameters: account for deprecation of overloaded pg_*() signatures (RFC)
> - PgSQL:
>  . Calling pg_fetch_result() with 2 arguments is deprecated. Use the
>    3-parameter signature with a null $row parameter instead.
>  . Calling pg_field_prtlen() with 2 arguments is deprecated. Use the
>    3-parameter signature with a null $row parameter instead.
>  . Calling pg_field_is_null() with 2 arguments is deprecated. Use the
>    3-parameter signature with a null $row parameter instead.

Note: when named parameters are used, the sniff will use the correct parameter name in the deprecation warning.
When positional parameters are used, the sniff will flag `$field` as missing, even though in reality (due to the secondary signature), the `$row` parameter is missing.

This could possible be clarified by adding an "alternative" key to the array and using the value of that in the error message, but that's for later.

Refs:
* https://wiki.php.net/rfc/deprecate_functions_with_overloaded_signatures#pg_fetch_result_pg_field_prtlen_and_pg_field_is_null
* php/php-src 12728
* https://github.com/php/php-src/commit/beaf1e814a3e8886ebcc59a050b5a5d6a0a46368
* https://www.php.net/pg_fetch_result
* https://www.php.net/pg_field_prtlen
* https://www.php.net/pg_field_is_null

Related to 1589