Composer: update the PHPUnit Polyfills to v 3.0 and allow for PHPUnit 11
A PHPUnit 11 compatible version of the PHPUnit Polyfills has been released and the `UtilityMethodTestCase` from PHPCSUtils is also fully compatible with PHPUnit 11, so let's start using it.

Mind: PHPUnit 11 throws (~1000) deprecation notices about the use of annotations instead of attributes.
On PHPUnit < 10.5.32 and PHPUnit < 11.3.3, these deprecation notices would fail the test run.

PHPUnit 10.5.32 and 11.3.3 introduce two new options for both the configuration file and as CLI arguments: `displayDetailsOnPhpunitDeprecations` and `failOnPhpunitDeprecation`, both of which will now default to `false`, which means the test runs will no longer fail, making the update viable.

These options have been added to the PHPUnit configuration for PHPUnit 10/11 to safeguard them against changes in the default value.

Includes updating the GitHub Actions workflows to use the correct PHPUnit config file based on the PHPUnit version (and allowing for PHPUnit 11).

Ref:
* https://github.com/Yoast/PHPUnit-Polyfills/releases/tag/3.0.0
* https://phpunit.de/announcements/phpunit-11.html
* sebastianbergmann/phpunit 5937