GH Actions/PHPUnit config: deal with removed `--coverage-cache` CLI option
The `--coverage-cache` flag was silently deprecated in PHPUnit 10 and has been removed in PHPUnit 11.

The `--cache-directory` flag replaces it and can also be set in the XML file.

This commit removes the outdated flag from the commands used in GH Actions and adds the attribute to the PHPUnit config file.

Ref:
* sebastianbergmann/phpunit 4599