PCRERegexTrait::getRegexModifiers(): bug fix - ignore text within brackets
When the regex pattern is retrieved via a function call, an unrelated - i.e. not part of the actual regex - text parameter may be passed.
When the regex pattern is retrieved from an array, the key may contain a text string. Again, this text string would not be part of the actual regex.

In both cases, any such text string should be ignored when collecting the regex parts to eventually determine if there are modifiers we can examine.

Fixed now by skipping over anything within brackets.

Includes tests.